+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## unreleased
 * Deprecate `MasterpassCard` payment method
 * Deprecate `AmexExpressCheckoutCard` payment method
+* Deprecate `deviceSessionId` and `fraudMerchantId` in `CreditCardRequest`, `CustomerRequest`, `PaymentMethodRequest`, and `TransactionRequest` classes
+* Fix issue where risk data was not included in `CustomerRequest` (fixes #85)
 
 ## 3.1.0
 * Add `RISK_THRESHOLD` to `GatewayRejectionReason`